Move GraphQL config to its own module
